+++ sources/node/src/logic/middleware/messages/chat_forward.js
@@ -23,16 +23,19 @@ class ChatForwardMessage extends Composer {
             return next();
         }
 
+        // try warn
         // delete message
-        // warn
-        await context.deleteMessage();
-        warn(
-            context,
-            this.database,
-            context.message.from.id,
-            1,
-            "Channel forward"
-        );
+        if (
+            warn(
+                context,
+                this.database,
+                context.message.from.id,
+                1,
+                "Channel forward"
+            ) > 0
+        ) {
+            await context.deleteMessage();
+        }
     }
 }
